chore: restructure packages by moving classes into domain folder
