+++ src/main/kotlin/racingcar/Printer.kt
@@ -1,5 +1,7 @@
 package racingcar
 
+import racingcar.domain.Car
+
 interface Printer {
     fun printCurrentCarPosition(car: List<Car>)
 }
