+++ src/main/kotlin/racingcar/domain/Engine.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 interface Engine {
     fun move() : Int
