+++ src/main/kotlin/racingcar/domain/RaceGame.kt
@@ -1,9 +1,12 @@
-package racingcar
+package racingcar.domain
+
+import racingcar.Printer
 
 class RaceGame(
     private val cars: List<Car>,
     private val roundNumber : Int,
-    private val resultPrinter: Printer) {
+    private val resultPrinter: Printer
+) {
 
     fun raceAllRound() : List<Car> {
         for (i in 1..roundNumber) {
