+++ src/main/kotlin/racingcar/domain/RandomEngine.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 import camp.nextstep.edu.missionutils.Randoms
