+++ src/test/kotlin/racingcar/RaceGameTest.kt
@@ -3,6 +3,9 @@ package racingcar
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import racingcar.domain.Car
+import racingcar.domain.Engine
+import racingcar.domain.RaceGame
 
 class RaceGameTest {
