+++ src/test/kotlin/racingcar/WinnerTest.kt
@@ -3,6 +3,8 @@ package racingcar
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import racingcar.domain.Car
+import racingcar.domain.Winner
 
 class WinnerTest {
